FIX: signed/unsigned comparison
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@3947 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Registration/cmtkEchoPlanarUnwarpFunctional.cxx
@@ -337,7 +337,7 @@ cmtk::EchoPlanarUnwarpFunctional
   const int phaseEncodeDirection = function.m_PhaseEncodeDirection;
 
   // reset gradient vector
-  for ( size_t i = g.getlowbound(); i <= g.gethighbound(); ++i )
+  for ( int i = g.getlowbound(); i <= g.gethighbound(); ++i )
     g(i) = 0.0;
 
   const UniformVolume& sourceImage = *(function.m_ImageGrid);
